Timer
General fixes for Timer support

+++ Libraries/timer.h
@@ -31,12 +31,12 @@
 #include "macros.h"
 
 void enable_TMR0(bool enable,char select_edge,char select_clock,bool enable_Prescaler,int set_Prescaler);
-void set_TMR0(int timer);
+void set_TMR0(uint8_t timer);
 uint8_t get_TMR0();
 void enable_TMR1(bool enable,char select_clock,int set_Prescaler,bool enable_sychronize);
-void set_TMR1(int time);
+void set_TMR1(uint16_t time);
 uint16_t get_TMR1();
 void enable_TMR2(bool enable,int set_Prescaler,int set_Postscaler);
-void set_TMR2(int time);
+void set_TMR2(uint8_t time);
 uint8_t get_TMR2();
 #endif
